packages/feishu/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quanttide-feishu"
version = "0.1.0"
description = "飞书知识库 Python SDK - 文档转换、JupyterBook 生成"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lark-oapi>=1.5.3",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
